[feat]TODO수정, 삭제 작성자 체크 추가

+++ src/todos/commands/handlers/delete-todo.handler.ts
@@ -6,6 +6,7 @@ import { DropTodosEvent } from '../../events/impl/drop-todos.event';
 import { DropTodoEvent } from '../../events/impl/drop-todo.event';
 import { DeleteTodoCommand } from '../impl/delete-todo.command';
 import { HttpService } from '@nestjs/axios';
+import { TodoPermissonError } from '../../error/todo-permisson.error';
 
 @CommandHandler(DeleteTodoCommand)
 export class DeleteTodoHandler implements ICommandHandler<DeleteTodoCommand> {
@@ -17,8 +18,11 @@ export class DeleteTodoHandler implements ICommandHandler<DeleteTodoCommand> {
   ) {}
 
   async execute(command: DeleteTodoCommand) {
-    const { todoId } = command;
+    const { userId, todoId } = command;
     const todo = await this.todoRepository.findOneBy({ todoId });
+    if (userId !== todo.userId) {
+      throw new TodoPermissonError();
+    }
     const date = todo.todoDate.toString();
 
     await this.todoRepository.delete(todoId);
